+++ include/cg_draw.h
@@ -50,6 +50,8 @@ void CG_Draw3DModel(
   vec3_t const origin,
   vec3_t const angles);
 
+void CG_DrawLinePitch(float angle, float pitch, float x, float w, float h, vec4_t const color);
+
 void CG_FillAngleYaw(float start, float end, float yaw, float y, float h, vec4_t const color);
 void CG_DrawLineYaw(float angle, float yaw, float y, float w, float h, vec4_t const color);
 void CG_DrawCharYaw(float angle, float yaw, float y, float w, float h, uint8_t ch, vec4_t const color);