+++ include/cg_local.h
@@ -187,8 +187,8 @@ typedef struct
   gameState_t gameState;    // gamestate from server
   glconfig_t  glconfig;     // rendering configuration
   float       screenXScale; // derived from glconfig
-  float       screenYScale;
-  float       screenXBias;
+  float       screenWidth;  // normalized/virtual screen width  (always SCREEN_WIDTH, i.e. 640)
+  float       screenHeight; // normalized/virtual screen height (depends on aspect ratio, e.g. 4:3 => 480, 16:9 => 360)
 
   int32_t levelStartTime;