+++ src/pitch.h
@@ -1,0 +1,10 @@
+#ifndef PITCH_H
+#define PITCH_H
+
+void init_pitch(void);
+
+void update_pitch(void);
+
+void draw_pitch(void);
+
+#endif // PITCH_H